'additing_dialog'

+++ src/scenes/dashboard/Dashboard.jsx
@@ -17,8 +17,9 @@ import InputLabel from '@material-ui/core/InputLabel';
 // import Input from '@material-ui/core/Input';
 import Input from '../../components/input/Input'
 import TextField from '@material-ui/core/TextField';
-
 import * as styled from './Dashboard.styled';
+import * as styledDialog from '../../components/dialog/AlertDialog.styles';
+import { AlertDialog } from "../../components/dialog/AlertDialog";
 import Task from './task/Task';
 import pushpin from '../../image/pushpin.svg';
 import low from '../../image/low.svg';
@@ -62,6 +63,8 @@ export class Dashboard extends Component {
             newComment: props.comment,
             priority: 'NOT_SPECIFIED',
             visibleAllDashboardTags: false,
+            visibleDelete: false,
+            visibleRestore: false,
         };
     }
 
@@ -92,6 +95,20 @@ export class Dashboard extends Component {
         this.setState({
             statePopup: false,
         });
+    };
+
+    showAlertDeleteDialog = () => {
+        const { visibleDelete } = this.state;
+        this.setState({
+            visibleDelete: !visibleDelete,
+        })
+    };
+
+    showAlertRestoreDialog = () => {
+        const { visibleRestore } = this.state;
+        this.setState({
+            visibleRestore: !visibleRestore,
+        })
     };
 
     handleUpdateTitle = (newValue) => {
@@ -147,7 +164,7 @@ export class Dashboard extends Component {
             tagTaskKeys,
         } = this.props;
         const {
-            valueNewTask, statePopup, stateComment, priority, visibleAllDashboardTags,
+            valueNewTask, statePopup, stateComment, priority, visibleAllDashboardTags, visibleDelete, visibleRestore,
         } = this.state;
 
         return ([
@@ -169,7 +186,7 @@ export class Dashboard extends Component {
                             />
                         ) : null
                     }
-                    <styled.Title
+                    <Input
                         onChange={this.handleUpdateTitle}
                         value={title}
                         onBlur={this.handleUpdateTitleSuccess}
@@ -232,19 +249,27 @@ export class Dashboard extends Component {
                                         </IconButton>
                                         <IconButton
                                             aria-label="trash"
-                                            onClick={() => actions.deleteDashboard({ id: idList })}
+                                            onClick={this.showAlertDeleteDialog}
                                             style={{ borderRadius: '40%', padding: '4px' }}
                                             alt="Delete this list"
                                         >
                                             <Delete />
                                         </IconButton>
+                                        <styledDialog.Dialog>
+                                            <AlertDialog
+                                                visible={visibleDelete}
+                                                onClose={this.showAlertDeleteDialog}
+                                                value='Do you want to delete this list?'
+                                                onConfirm={() => actions.deleteDashboard({ id: idList })}
+                                            />
+                                        </styledDialog.Dialog>
                                     </styled.IconContainer>
                                 )
                         ) : (
                             <styled.IconContainer>
                                 <IconButton
                                     aria-label="restore"
-                                    onClick={() => actionsBasket.restoreList({ id: idList })}
+                                    onClick={this.showAlertRestoreDialog}
                                     alt="Restore this list"
                                     style={{ borderRadius: '40%', padding: '4px' }}
                                 >
@@ -252,15 +277,29 @@ export class Dashboard extends Component {
                                         style={{ borderRadius: '40%' }}
                                     />
                                 </IconButton>
+                                <AlertDialog
+                                    visible={visibleRestore}
+                                    onClose={this.showAlertRestoreDialog}
+                                    value='Do you want to restore this list?'
+                                    onConfirm={() => actionsBasket.restoreList({ id: idList })}
+                                />
                                 <IconButton
                                     aria-label="trash"
-                                    onClick={() => actionsBasket.deletedList({ id: idList })}
+                                    onClick={this.showAlertDeleteDialog}
                                     style={{ borderRadius: '40%', padding: '4px' }}
                                     alt="Delete this list"
 
                                 >
                                     <Delete />
                                 </IconButton>
+                                <styledDialog.Dialog>
+                                    <AlertDialog
+                                        visible={visibleDelete}
+                                        onClose={this.showAlertDeleteDialog}
+                                        value='Do you want to delete this list forever?'
+                                        onConfirm={() => actionsBasket.deletedList({ id: idList })}
+                                    />
+                                </styledDialog.Dialog>
                             </styled.IconContainer>
                         )
                     }
